+++ packages/vibrant-components/src/lib/ContainedButton/ContainedButton.tsx
@@ -1,0 +1,65 @@
+import { Box, Text } from '@vibrant-ui/core';
+import { Icon } from '@vibrant-ui/icons';
+import { HStack } from '../HStack';
+import { Pressable } from '../Pressable';
+import { Spinner } from '../Spinner';
+import { withContainedButtonVariation } from './ContainedButtonProps';
+
+export const ContainedButton = withContainedButtonVariation(
+  ({
+    innerRef,
+    type = 'button',
+    IconComponent,
+    disclosure,
+    loading,
+    disabled,
+    onColor,
+    typography,
+    spinnerSize,
+    children,
+    ...restProps
+  }) => (
+    <Pressable
+      {...restProps}
+      ref={innerRef}
+      as="button"
+      buttonType={type}
+      interactions={['hover', 'focus', 'active']}
+      overlayColor={onColor}
+      borderRadiusLevel={1}
+      disabled={loading || disabled}
+    >
+      <>
+        <HStack as="span" alignVertical="center" alignHorizontal="center" opacity={loading ? 0 : 1}>
+          {IconComponent && (
+            <Box as="span" mx={2}>
+              <IconComponent size={16} />
+            </Box>
+          )}
+          <Text typography={typography} fontWeight="bold" mx={4}>
+            {children}
+          </Text>
+          {disclosure && (
+            <Box as="span" ml={2}>
+              <Icon.ArrowTriangleDown.Fill size={16} />
+            </Box>
+          )}
+        </HStack>
+        {loading && (
+          <HStack
+            as="span"
+            alignVertical="center"
+            alignHorizontal="center"
+            position="absolute"
+            top={0}
+            left={0}
+            right={0}
+            bottom={0}
+          >
+            <Spinner size={spinnerSize} />
+          </HStack>
+        )}
+      </>
+    </Pressable>
+  )
+);