+++ packages/vibrant-components/src/lib/ModalBottomSheet/ModalBottomSheet.spec.tsx
@@ -1,26 +1,10 @@
-import mediaQuery from 'css-mediaquery';
-import { waitFor } from '@testing-library/react';
+import { fireEvent, waitFor, waitForElementToBeRemoved } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { Box, PortalRootProvider } from '@vibrant-ui/core';
 import type { ReactRenderer } from '@vibrant-ui/utils/testing-web';
-import { createReactRenderer } from '@vibrant-ui/utils/testing-web';
+import { BrowserTesting, createReactRenderer } from '@vibrant-ui/utils/testing-web';
 import { Pressable } from '../Pressable';
 import { ModalBottomSheet } from './ModalBottomSheet';
-
-function createMatchMedia(width: number) {
-  return (query: string): MediaQueryList => ({
-    matches: mediaQuery.match(query, {
-      width,
-    }),
-    media: query,
-    onchange: null,
-    addListener: jest.fn(),
-    removeListener: jest.fn(),
-    addEventListener: jest.fn(),
-    removeEventListener: jest.fn(),
-    dispatchEvent: jest.fn(),
-  });
-}
 
 describe('<ModalBottomSheet />', () => {
   const { render } = createReactRenderer(children => <PortalRootProvider zIndex={1}>{children}</PortalRootProvider>);
@@ -65,12 +49,10 @@ describe('<ModalBottomSheet />', () => {
     });
 
     describe('after clicking the close button', () => {
-      beforeEach(done => {
-        waitFor(() => {
-          userEvent.click(renderer.getByRole('button', { name: 'Close' }));
-        }).then(() => {
-          setTimeout(done, 250);
-        });
+      beforeEach(async () => {
+        fireEvent.click(renderer.getByRole('button', { name: 'Close' }));
+
+        await waitForElementToBeRemoved(() => renderer.queryByRole('dialog'));
       });
 
       it('should call onClose function', () => {
@@ -113,7 +95,7 @@ describe('<ModalBottomSheet />', () => {
     });
   });
 
-  describe('when primaryButtonText and onPrimaryButtonClick are provided', () => {
+  describe('when primaryButtonOptions are provided', () => {
     let mockOnPrimaryButtonClick: jest.Mock<any, any>;
 
     beforeEach(() => {
@@ -122,10 +104,12 @@ describe('<ModalBottomSheet />', () => {
       renderer = render(
         <ModalBottomSheet
           defaultOpen={true}
-          primaryButtonText="primaryButtonText"
-          onPrimaryButtonClick={mockOnPrimaryButtonClick}
-          secondaryButtonText={undefined}
-          subButtonText={undefined}
+          primaryButtonOptions={{
+            title: 'primaryButtonText',
+            onClick: mockOnPrimaryButtonClick,
+          }}
+          secondaryButtonOptions={undefined}
+          subButtonOptions={undefined}
           renderOpener={({ open }) => <Pressable onClick={open} />}
         />
       );
@@ -146,7 +130,7 @@ describe('<ModalBottomSheet />', () => {
     });
   });
 
-  describe('when primaryButtonText, onPrimaryButtonClick, secondaryButtonText and onSecondaryButtonClick are provided', () => {
+  describe('when primaryButtonOptions, secondaryButtonOptions are provided', () => {
     let mockOnPrimaryButtonClick: jest.Mock<any, any>;
     let mockOnSecondaryButtonClick: jest.Mock<any, any>;
 
@@ -158,11 +142,15 @@ describe('<ModalBottomSheet />', () => {
       renderer = render(
         <ModalBottomSheet
           defaultOpen={true}
-          primaryButtonText="primaryButtonText"
-          onPrimaryButtonClick={mockOnPrimaryButtonClick}
-          secondaryButtonText="secondaryButtonText"
-          onSecondaryButtonClick={mockOnSecondaryButtonClick}
-          subButtonText={undefined}
+          primaryButtonOptions={{
+            title: 'primaryButtonText',
+            onClick: mockOnPrimaryButtonClick,
+          }}
+          secondaryButtonOptions={{
+            title: 'secondaryButtonText',
+            onClick: mockOnSecondaryButtonClick,
+          }}
+          subButtonOptions={undefined}
           renderOpener={({ open }) => <Pressable onClick={open} />}
         />
       );
@@ -197,7 +185,7 @@ describe('<ModalBottomSheet />', () => {
     });
   });
 
-  describe('when primaryButtonText, onPrimaryButtonClick, subButtonText and onSubButtonClick are provided', () => {
+  describe('when primaryButtonOptions, subButtonOptions are provided', () => {
     let mockOnPrimaryButtonClick: jest.Mock<any, any>;
     let mockOnSubButtonClick: jest.Mock<any, any>;
 
@@ -209,11 +197,15 @@ describe('<ModalBottomSheet />', () => {
       renderer = render(
         <ModalBottomSheet
           defaultOpen={true}
-          primaryButtonText="primaryButtonText"
-          onPrimaryButtonClick={mockOnPrimaryButtonClick}
-          secondaryButtonText={undefined}
-          subButtonText="subButtonText"
-          onSubButtonClick={mockOnSubButtonClick}
+          primaryButtonOptions={{
+            title: 'primaryButtonText',
+            onClick: mockOnPrimaryButtonClick,
+          }}
+          secondaryButtonOptions={undefined}
+          subButtonOptions={{
+            title: 'subButtonText',
+            onClick: mockOnSubButtonClick,
+          }}
           renderOpener={({ open }) => <Pressable onClick={open} />}
         />
       );
@@ -249,24 +241,28 @@ describe('<ModalBottomSheet />', () => {
   });
 
   describe('given mobile viewport width', () => {
-    beforeEach(() => {
-      window.matchMedia = createMatchMedia(639);
+    beforeEach(async () => {
+      await BrowserTesting.resize(639, 768);
+
+      renderer = render(<ModalBottomSheet open={true} renderContents={() => null} />);
     });
 
-    it('match snapshot', () => {
-      renderer = render(<ModalBottomSheet defaultOpen={true} renderContents={() => null} renderOpener={() => null} />);
+    it('match snapshot', async () => {
+      await waitFor(() => expect(renderer.getByRole('dialog').style.transform).toBe(''));
 
       expect(renderer.container).toMatchSnapshot();
     });
   });
 
   describe('given viewport width larger than mobile', () => {
-    beforeEach(() => {
-      window.matchMedia = createMatchMedia(1024);
+    beforeEach(async () => {
+      await BrowserTesting.resize(1024, 768);
+
+      renderer = render(<ModalBottomSheet open={true} renderContents={() => null} />);
     });
 
-    it('match snapshot', () => {
-      renderer = render(<ModalBottomSheet defaultOpen={true} renderContents={() => null} renderOpener={() => null} />);
+    it('match snapshot', async () => {
+      await waitFor(() => expect(renderer.getByRole('dialog').style.opacity).toBe('1'));
 
       expect(renderer.container).toMatchSnapshot();
     });