+++ packages/vibrant-core/src/lib/OverlayBox/OverlayBox.tsx
@@ -1,3 +1,4 @@
+import type { MouseEventHandler } from 'react';
 import { useCallback, useEffect, useState } from 'react';
 import type { Rect } from '@vibrant-ui/utils';
 import { getElementRect } from '@vibrant-ui/utils';
@@ -5,6 +6,7 @@ import { Box } from '../Box';
 import { PortalBox } from '../PortalBox';
 import { PressableBox } from '../PressableBox';
 import { useLockBodyScroll } from '../useLockBodyScroll';
+import { useWindowDimensions } from '../useWindowDimensions';
 import { withOverlayBoxVariation } from './OverlayBoxProps';
 
 export const OverlayBox = withOverlayBoxVariation(
@@ -13,19 +15,30 @@ export const OverlayBox = withOverlayBoxVariation(
 
     useLockBodyScroll(open);
 
+    const { width, height } = useWindowDimensions();
+
     useEffect(() => {
       if (!open) {
         return;
       }
 
-      getElementRect(targetRef.current).then(rect => setTargetRect(rect));
-    }, [open, targetRef]);
+      getElementRect(targetRef.current).then(rect => {
+        setTargetRect(rect);
+      });
+    }, [open, targetRef, width, height]);
 
-    const handleTargetTouchEnd = useCallback(() => {
-      onDismiss?.();
+    const handleTargetClick: MouseEventHandler = useCallback(
+      event => {
+        if (event.target !== event.currentTarget) {
+          return;
+        }
 
-      setTargetRect(null);
-    }, [onDismiss]);
+        onDismiss?.();
+
+        setTargetRect(null);
+      },
+      [onDismiss]
+    );
 
     if (!targetRect || !open) {
       return null;
@@ -34,17 +47,18 @@ export const OverlayBox = withOverlayBoxVariation(
     return (
       <PortalBox zIndex={zIndex} top={0} right={0} bottom={0} left={0}>
         <PressableBox position="absolute" width="100%" height="100%" onClick={() => onDismiss?.()} />
-        <PressableBox
+        <Box
+          as="div"
           top={targetRect.y}
           left={targetRect.x}
           width={targetRect.width}
           height={targetRect.height}
-          onClick={handleTargetTouchEnd}
+          onClick={handleTargetClick}
         >
           <Box position="absolute" ref={innerRef} {...boxProps}>
             {children}
           </Box>
-        </PressableBox>
+        </Box>
       </PortalBox>
     );
   }