+++ packages/vibrant-core/src/lib/props/interaction/type.ts
@@ -2,4 +2,5 @@ import type { ResponsiveValue } from '../../../types';
 
 export type InteractionSystemProps = {
   cursor?: ResponsiveValue<'default' | 'pointer' | 'text'>;
+  pointerEvents?: ResponsiveValue<'auto' | 'none'>;
 };