+++ packages/vibrant-core/src/types/ReactTextChildren.ts
@@ -1,0 +1,3 @@
+import type { ReactTextChild } from './ReactTextChild';
+
+export type ReactTextChildren = ReactTextChild | ReactTextChildren[];