+++ packages/utils-eslint-config/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+### [0.24.2](https://github.com/pedaling/opensource/compare/utils-eslint-config-0.24.1...utils-eslint-config-0.24.2) (2022-10-31)
+
+### [0.24.1](https://github.com/pedaling/opensource/compare/utils-eslint-config-0.24.0...utils-eslint-config-0.24.1) (2022-10-31)
+
+## [0.24.0](https://github.com/pedaling/opensource/compare/utils-eslint-config-0.23.0...utils-eslint-config-0.24.0) (2022-10-31)
+
 ## [0.23.0](https://github.com/pedaling/opensource/compare/utils-eslint-config-0.22.0...utils-eslint-config-0.23.0) (2022-10-27)
 
 ## [0.22.0](https://github.com/pedaling/opensource/compare/utils-eslint-config-0.21.2...utils-eslint-config-0.22.0) (2022-10-24)